requirements.txt
PyPDF2==1.28.4
prettytable==3.9.0
pillow==10.0.1